cmd Color - fully working (tested thouroughly) / cd slighting working but implemented  within terminal module
implemented cmdColor fully - three additions outside of cmdColor to make it work - (1)#setupHTML: #eTextbox (the tag where all text goes) i used setAttribute so that this tag would have a class. (2) inside style.css, I referenced the previously created tag rather than .pre, .p . This is because the latter version was making the stylesheets textcolor take precedence over the new text-color from cmdcolor. (3)#setupHTML: #eTerminal background color set to the same as it is in style.css (black) through inline css. This is to prevent the edge case where the user inputs "color 0", in which the foreground and background would both be 0.c

(i did not realize that nameTree has getCurrentPath() so we should probably use cd using that in the future) began implementation of CD - created directoryStack variable in terminal and changed awaitCommand() variable to print directoryStack(). ENCOUNTERED BUG: cd Program File does not work, likely a result of our filePath class not working when there is a space in the directory name.